pyproject.toml
[project]
name = "model-context-protocol"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ai-models>=0.7.4",
    "anthropic>=0.55.0",
    "asyncio>=3.4.3",
    "boto3>=1.38.46",
    "httpx>=0.28.1",
    "mcp[cli]>=1.10.1",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.1",
]
